requirements.txt
Flask
mysql-connector-python
beautifulsoup4
requests
